# docs/swagger.yaml
components:
  schemas:
    domain.Address:
      properties:
        district:
          type: string
        province:
          type: string
        subdistrict:
          type: string
        zipcode:
          type: string
      required:
      - district
      - province
      - subdistrict
      - zipcode
      type: object
    domain.CheckoutStatus:
      type: string
      x-enum-varnames:
      - StatusOpen
      - StatusComplete
      - StatusExpired
    domain.Dorm:
      properties:
        address:
          $ref: '#/components/schemas/domain.Address'
        bathrooms:
          minimum: 0
          type: integer
        bedrooms:
          minimum: 0
          type: integer
        createAt:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        owner:
          $ref: '#/components/schemas/domain.User'
        ownerId:
          type: string
        price:
          type: number
        rating:
          maximum: 5
          minimum: 0
          type: number
        size:
          type: number
        updateAt:
          type: string
      required:
      - address
      - bathrooms
      - bedrooms
      - name
      - ownerId
      - price
      - size
      type: object
    domain.LeasingHistory:
      properties:
        dorm:
          $ref: '#/components/schemas/domain.Dorm'
        dorm_id:
          type: string
        end:
          type: string
        id:
          type: string
        lessee:
          $ref: '#/components/schemas/domain.User'
        lessee_id:
          type: string
        orders:
          items:
            $ref: '#/components/schemas/domain.Order'
          type: array
          uniqueItems: false
        start:
          type: string
      type: object
    domain.Lifestyle:
      type: string
      x-enum-varnames:
      - Active
      - Creative
      - Social
      - Relaxed
      - Football
      - Basketball
      - Tennis
      - Swimming
      - Running
      - Cycling
      - Badminton
      - Yoga
      - GymAndFitness
      - Music
      - Dancing
      - Photography
      - Painting
      - Gaming
      - Reading
      - Writing
      - DIYAndCrafting
      - Cooking
      - Extrovert
      - Introvert
      - NightOwl
      - EarlyBird
      - Traveler
      - Backpacker
      - NatureLover
      - Camping
      - BeachLover
      - DogLover
      - CatLover
      - Freelancer
      - Entrepreneur
      - OfficeWorker
      - RemoteWorker
      - Student
      - SelfEmployed
    domain.Order:
      properties:
        createAt:
          type: string
        id:
          type: string
        leasingHistory:
          $ref: '#/components/schemas/domain.LeasingHistory'
        leasingHistoryID:
          type: string
        paidTransaction:
          $ref: '#/components/schemas/domain.Transaction'
        paidTransactionID:
          type: string
        price:
          type: integer
        transactions:
          items:
            $ref: '#/components/schemas/domain.Transaction'
          type: array
          uniqueItems: false
        type:
          $ref: '#/components/schemas/domain.OrderType'
        updateAt:
          type: string
      type: object
    domain.OrderType:
      type: string
      x-enum-varnames:
      - InsuranceOrderType
      - MonthlyBillOrderType
    domain.Role:
      type: string
      x-enum-varnames:
      - AdminRole
      - LesseeRole
      - LessorRole
    domain.Transaction:
      properties:
        createAt:
          type: string
        id:
          type: string
        price:
          type: integer
        status:
          $ref: '#/components/schemas/domain.CheckoutStatus'
        updateAt:
          type: string
      type: object
    domain.User:
      properties:
        birthDate:
          type: string
        createAt:
          type: string
        email:
          type: string
        filledPersonalInfo:
          type: boolean
        firstname:
          type: string
        gender:
          type: string
        id:
          type: string
        isStudentVerified:
          type: boolean
        isVerified:
          type: boolean
        lastname:
          type: string
        lifestyles:
          items:
            $ref: '#/components/schemas/domain.Lifestyle'
          type: array
          uniqueItems: false
        nationalID:
          type: string
        phoneNumber:
          type: string
        role:
          $ref: '#/components/schemas/domain.Role'
        studentEvidence:
          description: studentEvidence
          type: string
        updateAt:
          type: string
        username:
          type: string
      required:
      - email
      - username
      type: object
...
    dto.DormRequestBody:
      properties:
        address:
          properties:
            district:
              type: string
            province:
              type: string
            subdistrict:
              type: string
            zipcode:
              type: string
          required:
          - district
          - province
          - subdistrict
          - zipcode
          type: object
        bathrooms:
          minimum: 0
          type: integer
        bedrooms:
          minimum: 0
          type: integer
        description:
          type: string
        name:
          type: string
        price:
          type: number
        size:
          type: number
      required:
      - address
      - bathrooms
      - bedrooms
      - name
      - price
      - size
      type: object
    dto.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    dto.Pagination:
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      type: object
    dto.PaginationResponse-domain_Dorm:
      properties:
        data:
          items:
            $ref: '#/components/schemas/domain.Dorm'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/dto.Pagination'
      type: object
    dto.PaginationResponse-domain_LeasingHistory:
      properties:
        data:
          items:
            $ref: '#/components/schemas/domain.LeasingHistory'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/dto.Pagination'
      type: object
...
    dto.SuccessResponse-domain_Dorm:
      properties:
        data:
          $ref: '#/components/schemas/domain.Dorm'
      type: object
    dto.SuccessResponse-domain_LeasingHistory:
      properties:
        data:
          $ref: '#/components/schemas/domain.LeasingHistory'
      type: object
...
  securitySchemes:
    Bearer:
      description: Bearer token authentication
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: This is the API for the Condormhub project.
  title: Condormhub API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /dorms:
    get:
      description: Retrieve a list of all dorms
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.PaginationResponse-domain_Dorm'
          description: All dorms retrieved successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Failed to retrieve dorms
      summary: Get all dorms
      tags:
      - dorms
    post:
      description: Add a new room to the database with the given details
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.DormRequestBody'
        description: Dorm information
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_Dorm'
          description: Dorm successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Your request is invalid
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Failed to save dorm
      security:
      - Bearer: []
      summary: Create a new dorm
      tags:
      - dorms
  /dorms/{id}:
    delete:
      description: Removes a dorm from the database based on the give ID
      parameters:
      - description: DormID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_Dorm'
          description: Dorm successfully deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Incorrect UUID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Dorm not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Failed to delete dorm
      security:
      - Bearer: []
      summary: Delete a dorm
      tags:
      - dorms
    get:
      description: Retrieve a specific dorm based on its ID
      parameters:
      - description: DormID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_Dorm'
          description: Dorm data successfully retrieved
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Incorrect UUID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Dorm not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Server failed to retrieve dorm
      summary: Get a dorm by ID
      tags:
      - dorms
    patch:
      description: Modifies an existing room's details based on the given ID
      parameters:
      - description: DormID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.DormRequestBody'
        description: Updated Room Data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_Dorm'
          description: Dorm data updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Invalid Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Dorm not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Server failed to update dorm
      security:
      - Bearer: []
      summary: Update an existing dorm
      tags:
      - dorms
  /history/{id}:
    delete:
      description: Delete a leasing history in the database
      parameters:
      - description: LeasingHistoryId
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Incorrect UUID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: leasing history not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Can not parse UUID or Failed to delete leasing history
      security:
      - Bearer: []
      summary: Delete a leasing history
      tags:
      - history
    patch:
      description: Delete a leasing history in the database
      parameters:
      - description: LeasingHistoryId
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Set end timestamp successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Incorrect UUID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: leasing history not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Can not parse UUID or Failed to update leasing history
      security:
      - Bearer: []
      summary: Delete a leasing history
      tags:
      - history
    post:
      description: Add a new leasing history to the database
      parameters:
      - description: DormID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_LeasingHistory'
          description: Dorm successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Incorrect UUID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Dorm not found or leasing history not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Can not parse UUID or failed to save leasing history to database
      security:
      - Bearer: []
      summary: Create a new leasing history
      tags:
      - history
  /history/bydorm/{id}:
    get:
      description: Retrieve a list of all leasing history by userid
      parameters:
      - description: DormID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Number of history to be retirved
        in: query
        name: limit
        required: true
        schema:
          type: string
      - description: Page to retrive
        in: query
        name: page
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.PaginationResponse-domain_LeasingHistory'
          description: Retrive history successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Incorrect UUID format or limit parameter is incorrect or page
            parameter is incorrect or page exceeded
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: leasing history not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Can not parse UUID
      security:
      - Bearer: []
      summary: Get all leasing history by userid
      tags:
      - history
...